# Improve PoS performance