Found a few more optimizations that speed up both proving and verification.

Before:
```
chia/table/parallel/8x  time:   [1.2842 s 1.2958 s 1.3070 s]
chia/verification       time:   [11.420 µs 11.438 µs 11.461 µs]
```

After:
```
chia/table/parallel/8x  time:   [1.0766 s 1.0850 s 1.0932 s]
chia/verification       time:   [11.215 µs 11.220 µs 11.226 µs]
```

Each commit introduces a non-negligible performance improvement.

Proving improvement is thanks to simpler code that is very favorable for modern CPUs with very long internal pipelines.

### Code contributor checklist:
* [x] I have read, understood and followed [contributing guide](https://github.com/autonomys/subspace/blob/main/CONTRIBUTING.md)