@nazar-pc 2024-09-10
It is indeed platform-specific to a degree. While ARM CPUs are not quite as RISC as some other RISC architectures, it does have pipelining and various tricks, but modern x86-64 CISC CPUs have huge pipelines and can do a lot of heavy lifting if you feed them well with data. So it is totally within expectations that ARM is not affected as much if at all.

You might find https://www.youtube.com/watch?v=wGSSUSeaLgA interesting (from C++Now 2024), it explains a lot of nuances with modern CPUs, especially large server ones.