@nazar-pc 2024-09-10
On older x86-64 CPUs the effect will also be much smaller if any at all. This optimization is specifically for modern CPUs and the trend where they are going. Even M1 vs M4 Apple Silicon might have a significantly different results with these changes.